+++ Backend/Repositories/InterestRepo/IInterestRepository.cs
@@ -6,12 +6,12 @@ namespace Backend.Repositories.InterestRepo
 {
     public interface IInterestRepository
     {
-        Interests GetInterestById(int id);
+        Interests GetInterestById(long id);
         Interests GetInterestByName(string name);
         Interests CreateInterest(Interests interest);
         Interests UpdateInterest(Interests interest);
         void DeleteInterest(Interests interest);
-        List<Interests> GetStudentsInterests(int studentId);
-        List<Interests> UpdateStudentInterests(int studentId, List<Interests> interests);
+        List<Interests> GetStudentsInterests(long studentId);
+        List<Interests> UpdateStudentInterests(long studentId, List<Interests> interests);
     }
 }
